PHOENIX-368 View Demand Collection and Balance (DCB) of a Property

+++ egov/egov-demand/src/main/java/org/egov/demand/dao/EgDemandDao.java
@@ -45,7 +45,7 @@ import org.egov.demand.model.EgDemand;
 
 public interface EgDemandDao {
 	
-	public EgDemand findById(Integer id, boolean lock);
+	public EgDemand findById(Long id, boolean lock);
 
 	public List<EgDemand> findAll();
